finalizacion api

+++ src/main/java/com/Artisan/repository/CategoryRepository.java
@@ -3,6 +3,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.Artisan.entities.Category;
 
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-	
-}
+public interface CategoryRepository extends JpaRepository<Category, Integer> {}
